Create sim client
